Removed AntiVPN, fixed bugs, and added checks
- Hand (D): if the ray does not collide with the victim's boundingBox, it will not be used to check for block collisions.
- Increased the subtraction verbose for Velocity (B). It will prevent false positives and it still flags very quickly for velocity. Check if it needs to be increased even more.
- Implemented check to prevent false positives from BadPackets (O) when attacking block. Also commonly found on 1.12.2 (thx Smsgamer).
- Decreased the EvictingList size to 35 for Timer to allow for better performance since 50 is a little much.
- Changed Kauri#unload to Kauri#unload(boolean). When set true, it will save the database asynchronously. Only ment for when the server is not shutting down!
- Added Kauri isPaper field to fix certain PaperSpigot bugs. May move to Atlas eventually.
- Fixed disabler caused by the exploit of PacketPlayInBlockPlace. Kind of like a Mineplex fly from months ago.
- Removed AntiVPN and will replace with a hook into the KauriVPN plugin soon to be released on Spigot.
- Fixed false positive in Autoclicker (G) by checking for lag using timeStamps.
- Added an unblock method to patch autoblocks that just never unblock (vanilla autoblock).
- Added new Aim (K) vape check [elevated].
- Changed a few values to fix false positives and improve detection.
- Added new Aim (I) head snap angle check (by Itz_Lucky).
- Changed gcd to double in Aim (C) and Aim (B). Allows for much more accurate, faster, and better detection.
- Removed VanillaUtils.

+++ Impl/src/main/java/dev/brighten/anticheat/check/impl/combat/aim/AimC.java
@@ -12,12 +12,12 @@ import dev.brighten.api.check.CheckType;
         checkType = CheckType.AIM, punishVL = 100)
 public class AimC extends Check {
 
-    private float lastGcd;
+    private double lastGcd;
     @Packet
     public void onFlying(WrappedInFlyingPacket packet) {
         if(!packet.isLook()) return;
 
-        float gcd = (data.playerInfo.yawGCD / MovementProcessor.offset);
+        double gcd = (data.playerInfo.yawGCD / MovementProcessor.offset);
 
         if(MathUtils.getDelta(gcd, lastGcd) > 1E-4) {
             vl++;
